feat : 품의서 승인 / 반려 // 삭제 // 목록 조회 / 상세 조회 기능 구현

+++ src/main/java/com/harusari/chainware/requisition/command/application/dto/request/RequisitionItemRequest.java
@@ -4,7 +4,6 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 
 @Getter
 public class RequisitionItemRequest {
@@ -19,5 +18,5 @@ public class RequisitionItemRequest {
     private int quantity;
 
     @NotNull(message = "단가는 필수입니다.")
-    private BigDecimal unitPrice;
+    private Long unitPrice;
 }
